pad: break RDL router into several files
