+++ src/pad/src/RDLRouter.h
@@ -83,105 +83,8 @@ struct RouteTarget
   odb::dbTechLayer* layer;
 };
 
-class RDLGui : public gui::Renderer
-{
- public:
-  RDLGui();
-  ~RDLGui() override;
-
-  void setRouter(RDLRouter* router);
-
-  void drawObjects(gui::Painter& painter) override;
-
-  const char* getDisplayControlGroupName() override { return "RDL Router"; }
-
-  void pause(bool timeout) const;
-
- private:
-  RDLRouter* router_ = nullptr;
-
-  static constexpr const char* draw_vertex_ = "Vertices";
-  static constexpr const char* draw_edge_ = "Edges";
-  static constexpr const char* draw_obs_ = "Obstructions";
-  static constexpr const char* draw_fly_wires_ = "Routing fly wires";
-  static constexpr const char* draw_targets_ = "Targets";
-  static constexpr const char* draw_routes_ = "Routes";
-  static constexpr const char* draw_route_obstructions_ = "Route obstructions";
-
-  static constexpr int gui_timeout_ = 100;
-};
-
-class RDLRoute
-{
- public:
-  RDLRoute(odb::dbITerm* source,
-           const std::vector<odb::dbITerm*>& dests,
-           const std::map<grid_vertex, odb::Point>& vertex_point_map);
-
-  void setRoute(
-      const std::vector<grid_vertex>& vertex,
-      const std::set<std::tuple<odb::Point, odb::Point, float>>& removed_edges,
-      const RouteTarget* source,
-      const RouteTarget* target);
-  void resetRoute();
-
-  bool isRouted() const { return !route_vertex_.empty(); }
-  bool isFailed() const
-  {
-    return !route_pending_ && !isRouted() && !hasNextTerminal();
-  }
-  bool allowRipup(int other_priority) const
-  {
-    return isRouted() && other_priority >= priority_;
-  }
-
-  void markRouting() { route_pending_ = false; }
-
-  int getPriority() const { return priority_; }
-  odb::dbITerm* getTerminal() const { return iterm_; }
-  odb::dbNet* getNet() const { return iterm_->getNet(); }
-
-  const std::vector<odb::dbITerm*>& getTerminals() const { return terminals_; }
-
-  void increasePriority() { priority_++; }
-
-  bool hasNextTerminal() const { return next_ != terminals_.end(); }
-  odb::dbITerm* getNextTerminal();
-  odb::dbITerm* peakNextTerminal() const;
-  void moveNextTerminalToEnd() { next_ = terminals_.end(); }
-
-  bool compare(const std::shared_ptr<RDLRoute>& other) const;
-
-  const std::vector<grid_vertex>& getRouteVerticies() const
-  {
-    return route_vertex_;
-  }
-  const std::set<std::tuple<odb::Point, odb::Point, float>>& getRouteEdges()
-      const
-  {
-    return route_edges_;
-  }
-  const RouteTarget* getRouteTargetSource() const { return route_source_; }
-  const RouteTarget* getRouteTargetDestination() const { return route_dest_; }
-
-  bool isIntersecting(RDLRoute* other, int extent) const;
-
- private:
-  odb::dbITerm* iterm_;
-  int priority_;
-
-  bool route_pending_;
-
-  const std::map<grid_vertex, odb::Point>& vertex_point_map_;
-
-  std::vector<odb::dbITerm*> terminals_;
-  std::vector<odb::dbITerm*>::iterator next_;
-
-  std::vector<grid_vertex> route_vertex_;
-  std::set<std::tuple<odb::Point, odb::Point, float>> route_edges_;
-  const RouteTarget* route_source_;
-  const RouteTarget* route_dest_;
-};
+class RDLGui;
+class RDLRoute;
 
 class RDLRouter
 {
